chore: remove historical satellite layer (ECO-478)
Drop the Planet/NICFI monthly satellite basemap from the Green Globe map.
The hosted credits are exhausted and the license is no longer valid, so
the toggle now represents dead UI that could fire failed tile requests.

- Remove the LayersOverlay control, map hook, source/layer generators,
  and the /api/planet-tiles Next.js proxy.
- Trim historicalSatellite state/action from the layers store, and the
  "historical-satellite" slot from the URL navigation state + sync hook.
- Relocate the surviving landcover source/layer helpers into their own
  sources-and-layers/landcover.ts so the remaining WorldCover toggle
  keeps working.
- Drop the historical-satellite step, step definitions, and
  /api/planet-tiles fixture from the Playwright/Cucumber suite.
- Remove the committed NICFI_API_KEY (the key should also be rotated
  out of band since it was previously in git history).
