+++ e2e/features/layers.feature
@@ -7,7 +7,5 @@ Feature: Layers
     And the project-specific layer "Canopy Plots" is visible
     When the visitor enables the landcover layer
     Then the current URL contains "layers-landcover=true"
-    When the visitor enables the historical satellite layer
-    Then the current URL contains a historical satellite month
     When the visitor enables the project-specific layer "Canopy Plots"
     Then the current URL contains "Canopy%20Plots"
